rag-system-onE-commerce/requirements.txt
# RAG System - All Requirements


# RAG Shopping System - Complete Requirements
# Advanced Information Retrieval Course
# Updated for Phase 3

# ============================================================================
# Core Data Processing (Phase 1)
# ============================================================================
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ============================================================================
# Vector Database & Embeddings (Phase 2)
# ============================================================================
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
# Note: If you have GPU, replace faiss-cpu with: faiss-gpu>=1.7.4
tqdm>=4.65.0

# ============================================================================
# Advanced Retrieval (Phase 3)
# ============================================================================
rank-bm25>=0.2.2              # BM25 keyword-based retrieval
nltk>=3.8.1                   # Natural language processing
# Cross-encoder for re-ranking (uses sentence-transformers, already installed)

# ============================================================================
# Utilities
# ============================================================================
scikit-learn>=1.3.0
pathlib2>=2.3.7

# ============================================================================
# Development & Testing (Optional but Recommended)
# ============================================================================
jupyter>=1.0.0
ipython>=8.12.0
ipywidgets>=8.0.0
pytest>=7.0.0

# ============================================================================
# Future Phases (Will be added later)
# ============================================================================
# Phase 4 (Generation): openai, anthropic, langchain, etc.
# Phase 5 (UI): streamlit, gradio, flask, etc.
# Phase 6 (Self-Learning): Additional ML libraries